% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information_criteria.R
\name{BIC}
\alias{BIC}
\title{Computes BIC for ordinary least squares}
\usage{
BIC(obs_list, crit_value, param_nb)
}
\arguments{
\item{obs_list}{List of observed values to use for parameter estimation.
A \verb{named list} (names = situations names) of data.frame containing
one column named Date with the dates (Date or POSIXct format) of the different observations
and one column per observed variables with either the measured values or NA, if
the variable is not observed at the given date.
See details section for more information on the list of observations actually
used during the parameter estimation process.}

\item{crit_value}{Final value of the estimated criterion}

\item{param_nb}{Number of estimated parameters}
}
\value{
Value of the BIC criterion for ordinary least squares.
If called without arguments, returns a named list with element "name"
containing the name of the function and "species" containing
"Information criterion"
}
\description{
Computes BIC for ordinary least squares
}
