% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{bonsai_bio_wrapper}
\alias{bonsai_bio_wrapper}
\title{MyModel wrapper for CroptimizR}
\usage{
bonsai_bio_wrapper(model_options, param_values)
}
\arguments{
\item{model_options}{List containing any information needed to run the model.There
are 3 elements : path for finding the file, the beginning day, the end.}

\item{param_values}{Named 3D array that contains the value(s) and names of the
parameters to force for each situation to simulate. This array contains the different
parameters values (first dimension) for the different parameters (second dimension)
and for the different situations (third dimension).}
}
\value{
A list containing simulated values (\code{sim_list}: a vector of list (one
element per values of parameters) containing data.frames of simulated output values
for each simulated situation) and an error code (\code{error}) indicating if at least
one simulation ended with an error.
}
\description{
This function runs my crop model and force it with the values
of the parameters defined in the param_values argument. It returns
the values of the simulated outputs.
}
\keyword{internal}
