% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_protocol.R
\name{check_col_names}
\alias{check_col_names}
\title{Check the columns names in the protocol description file}
\usage{
check_col_names(protocol_file_path, expected_cols, cols, sheet)
}
\arguments{
\item{expected_cols}{List of expected column names for a given sheet}

\item{cols}{List of column names as read in the protocol description file for a given sheet}

\item{sheet}{Name of the sheet read in the protocol description file}
}
\value{
Stop if column names are not as expected
}
\description{
Check the columns names in the protocol description file
}
\details{
Check that the columns listed in \code{expected_cols} are included in columns listed in \code{cols}
for sheet \code{sheet}
}
\keyword{internal}
