% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsSim_consistency.R
\name{check_obsSim_consistency}
\alias{check_obsSim_consistency}
\title{Check the consistency of observation and simulation lists}
\usage{
check_obsSim_consistency(sim_list, obs_list)
}
\arguments{
\item{sim_list}{List of simulated values}

\item{obs_list}{List of observed values}
}
\description{
Check the consistency of observation and simulation lists
}
\details{
This function checks:
(i) that Date columns in obs and sim are either in Date or POSIXct formats
(ii) that variables in obs and sim list have same types and are numeric
}
\keyword{internal}
