% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_protocol.R
\name{check_protocol_content}
\alias{check_protocol_content}
\title{Check the content of a protocol description file}
\usage{
check_protocol_content(
  protocol_file_path,
  variables_df,
  varNames_corresp,
  simVar_units,
  transform_outputs,
  param_group,
  obsVar_group,
  sitNames_corresp
)
}
\arguments{
\item{protocol_file_path}{Path to the Excel file describing the AgMIP PhaseIV protocol.}

\item{variables_df}{to be defined ...}

\item{varNames_corresp}{to be defined ...}

\item{simVar_units}{to be defined ...}

\item{transform_outputs}{to be defined ...}

\item{param_group}{to be defined ...}

\item{obsVar_group}{to be defined ...}

\item{sitNames_corresp}{to be defined ...}
}
\value{
Stop if the content of the protocol description file is not as expected
}
\description{
Check the content of a protocol description file
}
\keyword{internal}
