% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_eq_const.R
\name{compute_eq_const}
\alias{compute_eq_const}
\title{Compute equality constraints on model wrapper input parameters}
\usage{
compute_eq_const(forced_param_values, param_values)
}
\arguments{
\item{forced_param_values}{Named vector or list, must contain the values (or
arithmetic expression, see details section) for the model parameters to force. The corresponding
values will be transferred to the model wrapper through its param_values argument
during the estimation process.
Should not include values for estimated parameters (i.e. parameters defined in
\code{param_info} argument), except if they are listed as candidate parameters (see
argument \code{candidate_param}).}

\item{param_values}{Named vector or tibble, value(s) of the estimated parameters
as provided by the parameter estimation algorithm (and possibly transformed in
main_crit function)}
}
\value{
Named vector or tibble (same shape as param_values) containing the values
for the model parameters to force.
}
\description{
Compute equality constraints on model wrapper input parameters
}
\keyword{internal}
