% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_simulations.R
\name{compute_simulations}
\alias{compute_simulations}
\title{Compute simulated variables by running the model wrapper and applying
the required transformations.}
\usage{
compute_simulations(
  model_function,
  model_options,
  param_values,
  situation,
  var_to_simulate,
  obs_list,
  transform_sim = NULL,
  transform_var = NULL,
  sit_var_dates_mask = NULL
)
}
\arguments{
\item{model_function}{Crop Model wrapper function to use.}

\item{model_options}{List of options for the Crop Model wrapper (see help of
the Crop Model wrapper function used).}

\item{param_values}{Named vector or tibble, value(s) of the parameters to
pass to the model wrapper}

\item{situation}{(optional) List of situations to take into account within obs_list.
situation = NULL means that all situations in obs_list will be used.}

\item{var_to_simulate}{(optional) List of variables for which the model wrapper must return
results.
By default the wrapper is asked to simulate only the observed variables. However,
it may be useful to simulate also other variables, typically when transform_sim
and/or transform_obs functions are used. Note however that it is
active only if the model_function used handles this argument.}

\item{obs_list}{List of observed values to use for parameter estimation.
A \verb{named list} (names = situations names) of data.frame containing
one column named Date with the dates (Date or POSIXct format) of the different observations
and one column per observed variables with either the measured values or NA, if
the variable is not observed at the given date.
See details section for more information on the list of observations actually
used during the parameter estimation procedure.}

\item{transform_sim}{User function for transforming simulations before each criterion
evaluation  (optional), see details section for more information.}

\item{transform_var}{Named vector of functions to apply both on simulated and
observed variables. \code{transform_var=c(var1=log, var2=sqrt)} will for example
apply log-transformation on simulated and observed values of variable var1,
and square-root transformation on values of variable var2.}

\item{sit_var_dates_mask}{List of situations: a named list
containing a mask for variables and dates for which simulated values
should be returned. Typically a list containing the observations to which
simulations should be compared.}
}
\value{
List of model and transformed results, in the cropr_simulation format.
}
\description{
Compute simulated variables by running the model wrapper and applying
the required transformations.
}
