% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_protocol_agmip.R
\name{compute_stat_per_group}
\alias{compute_stat_per_group}
\title{Compute statistics per group of variables}
\usage{
compute_stat_per_group(stat_col, step_var_map, stats_per_var)
}
\arguments{
\item{stat_col}{Name of the column of stats_per_var containing the statistics to compute}

\item{step_var_map}{A list mapping each variable to a list of variables used at the same step}

\item{stats_per_var}{A data frame containing statistics per variable}
}
\value{
A named vector containing for each variable the sum of the statistics for each group of variables
}
\description{
Compute statistics per group of variables
}
\keyword{internal}
