% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concentrated_wss.R
\name{concentrated_wss}
\alias{concentrated_wss}
\title{Compute concentrated version of weighted sum of squares from simulation and observation list. More details about this criterion are given in Wallach et al., 2011 (equation 5).}
\usage{
concentrated_wss(sim_list, obs_list)
}
\arguments{
\item{sim_list}{List of simulated variables}

\item{obs_list}{List of observed variables}
}
\value{
The value of the concentrated weighted sum of squares
}
\description{
Compute concentrated version of weighted sum of squares from simulation and observation list. More details about this criterion are given in Wallach et al., 2011 (equation 5).
}
\details{
\code{sim_list} and \code{obs_list} must have the same structure (i.e. same number of variables, dates, situations, ... use intersect_sim_obs for that).
}
