% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_param.R
\name{fill_step_info}
\alias{fill_step_info}
\title{Fill the step information}
\usage{
fill_step_info(step, mc)
}
\arguments{
\item{step}{(optional) List that describes the steps of the parameter estimation process (see details section).}

\item{mc}{A matched call object from \code{match.call()} in \code{estim_param},
containing the arguments explicitly provided by the user when calling \code{estim_param}.
This ensures that \code{fill_step_info} has access to the correct function call context.}
}
\value{
A list of step definitions provided by the user in the \code{estim_param} function call,
enriched with default values and additional arguments passed to \code{estim_param}.
The returned object is a list of lists, where each sublist contains the characteristics of a step.
}
\description{
Fill the step information
}
\keyword{internal}
