% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_obs.R
\name{filter_obs}
\alias{filter_obs}
\title{Filter observation list to exclude situations, variables or dates}
\usage{
filter_obs(obs_list, var_names = NULL, sit_names = NULL,
  dates = NULL, include = FALSE)
}
\arguments{
\item{obs_list}{List of observed values to use for parameter estimation
A \code{named list} (names = situations names) of data.frame containing
one column named Date with the dates (Date or POSIXct format) of the different observations
and one column per observed variables with either the measured values or NA, if
the variable is not observed at the given date.}

\item{var_names}{(optional, if not given all variables will be kept) Vector containing the names of the variables to include or exclude}

\item{sit_names}{(optional, if not given all situations will be kept) Vector containing the names of the situations to include or exclude}

\item{dates}{(optional, if not given all dates will be kept) Vector containing the dates (POSIXct format) to include or exclude}

\item{include}{(optional, FALSE by default) Flag indicating if the variables / situations / dates listed in inputs must be included (TRUE) or not (FALSE) in the resulting list}
}
\value{
obs_list List of filtered observed values (same format as \code{obs_list} input argument)
}
\description{
Filter observation list to exclude situations, variables or dates
}
\examples{

obs_list <- list(sit1=data.frame(Date=as.POSIXct(c("2009-11-30","2009-12-10")),
                                 var1=c(1.1,1.5),var2=c(NA,2.1)),
                 sit2=data.frame(Date=as.POSIXct(c("2009-11-30","2009-12-5")),
                                 var1=c(1.3,2)))

# Keep only var1
filter_obs(obs_list,var_names=c("var1"),include=TRUE)

# Exclude observations at date "2009-11-30"
filter_obs(obs_list,dates=as.POSIXct(c("2009-11-30")))


}
\seealso{
For more detail and examples, see the different vignettes in
\href{https://sticsrpacks.github.io/CroptimizR/}{CroptimizR website}
}
