% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_vignettes.R
\name{generate_vignette}
\alias{generate_vignette}
\title{Generate CroptimizR vignettes by assembling R scripts}
\usage{
generate_vignette(croptimizR_path = getwd(),
  vignettes = c("stics_simple", "stics_specific_varietal", "apsim",
  "dream"))
}
\arguments{
\item{croptimizR_path}{path to CroptimizR repository}

\item{vignettes}{vector of vignettes' names to generate (chose between those given by default)}
}
\value{
Generate R scripts and RMD file corresponding to the chosen vignettes
}
\description{
Generate CroptimizR vignettes by assembling R scripts
}
\keyword{internal}
