% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_agmip_protocol_functions.R
\name{get_agmip_protocol_template}
\alias{get_agmip_protocol_template}
\title{Get AgMIP protocol Excel template}
\usage{
get_agmip_protocol_template(path = ".", overwrite = FALSE)
}
\arguments{
\item{path}{Path where to copy the template file, or NULL to only return
the path to the file inside the package. Defaults to the current working
directory.}

\item{overwrite}{Logical. Overwrite existing file?}
}
\value{
The path to the template file (either inside the package or to the
copied file).
}
\description{
Returns the path to the AgMIP calibration protocol Excel template shipped
with CroptimizR, and copies it to a user-specified location.
}
