% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_obs_var.R
\name{get_obs_var}
\alias{get_obs_var}
\title{Get names of observed variables}
\usage{
get_obs_var(obs_list)
}
\arguments{
\item{obs_list}{List of observed values to use for parameter estimation.
A \verb{named list} (names = situations names) of data.frame containing
one column named Date with the dates (Date or POSIXct format) of the different observations
and one column per observed variables with either the measured values or NA, if
the variable is not observed at the given date.
See details section for more information on the list of observations actually
used during the parameter estimation procedure.}
}
\value{
Vector of names of observed variables
}
\description{
Get names of observed variables
}
