% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_reserved_keywords.R
\name{get_reserved_keywords}
\alias{get_reserved_keywords}
\title{Return reserved keywords that should not be used as (observed/simulated) variables names}
\usage{
get_reserved_keywords()
}
\value{
A vector including the names of the keywords
}
\description{
Return reserved keywords that should not be used as (observed/simulated) variables names
}
\keyword{internal}
