% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_sim_obs.R
\name{intersect_sim_obs}
\alias{intersect_sim_obs}
\title{Intersect simulation and observation lists}
\usage{
intersect_sim_obs(sim_list, obs_list)
}
\arguments{
\item{sim_list}{List of simulated values}

\item{obs_list}{List of observed values}
}
\value{
A list containing the new obs and sim lists having same situations,
variables, dates
}
\description{
Intersect simulation and observation lists
}
\keyword{internal}
