% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_data.R
\name{is.obs}
\alias{is.obs}
\title{Check format of observation list}
\usage{
is.obs(obs_list)
}
\arguments{
\item{obs_list}{List of observed values to check}
}
\value{
TRUE if obs_list has the good format, FALSE otherwise (+ warnings)
}
\description{
Check format of observation list
}
\examples{

obs_list <- list(sit1=data.frame(Date=as.POSIXct(c("2009-11-30","2009-12-10")),
                                 var1=c(1.1,1.5),var2=c(NA,2.1)),
                 sit2=data.frame(Date=as.POSIXct(c("2009-11-30","2009-12-5")),
                                 var1=c(1.3,2)))
CroptimizR:::is.obs(obs_list)

# Missing Date column
obs_list <- list(sit1=data.frame(var1=c(1.1,1.5),var2=c(NA,2.1)),
                 sit2=data.frame(Date=as.POSIXct(c("2009-11-30","2009-12-5")),
                                 var1=c(1.3,2)))
CroptimizR:::is.obs(obs_list)

# Bad Date format
obs_list <- list(sit1=data.frame(Date=c("2009-11-30","2009-12-10"),var1=c(1.1,1.5),
                                 var2=c(NA,2.1)),
                 sit2=data.frame(Date=c("2009-11-30","2009-12-5"),var1=c(1.3,2)))
CroptimizR:::is.obs(obs_list)

}
\keyword{internal}
