% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsSim_consistency.R
\name{is_sim_inf_or_na}
\alias{is_sim_inf_or_na}
\title{Check if sim contains inf/na values when there are observations}
\usage{
is_sim_inf_or_na(sim_list, obs_list, param_values)
}
\arguments{
\item{sim_list}{List of simulated values}

\item{obs_list}{List of observed values}

\item{param_values}{vector of input parameters used in the simulations}
}
\description{
Check if sim contains inf/na values when there are observations
}
\keyword{internal}
