% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{Likelihood}
\alias{Likelihood}
\alias{likelihood_ciidn}
\alias{likelihood_log_ciidn}
\alias{likelihood_ciidn_corr}
\title{Likelihood}
\usage{
likelihood_ciidn(sim_list, obs_list)

likelihood_log_ciidn(sim_list, obs_list)

likelihood_ciidn_corr(sim_list, obs_list)
}
\arguments{
\item{sim_list}{List of simulated variables}

\item{obs_list}{List of observed variables}
}
\value{
The value of the likelihood given the observed and simulated values of the variables.
}
\description{
Provide several likelihood to estimate parameters using bayesian methods.
}
\details{
The following likelihood are proposed (\href{https://sticsrpacks.github.io/CroptimizR/reference/likelihood.html}{see html version} for a better rendering of equations):
\itemize{
\item \code{likelihood_ciidn}: concentrated version of iid normal likelihood
\deqn{ \prod_{j} {\sum_{i,k} [Y_{ijk}-f_{jk}(X_i;\theta)]^2 )}^{-(n_j/2+2)} }
where \eqn{ Y_{ijk} } is the observed value for the \eqn{k^{th}} time point of the \eqn{j^{th}} variable in the \eqn{i^{th}}
situation,
\eqn{ f_{jk}(X_i;\theta) } the corresponding model prediction, and \eqn{n_j} the number of measurements of variable \eqn{j}. \cr
More details about this criterion are given in Wallach et al. (2011), equation 5.
\item \code{likelihood_log_ciidn}: log transformation of concentrated version of iid normal likelihood \cr
}
\code{sim_list} and \code{obs_list} must have the same structure (i.e. same number of variables, dates, situations, ... use internal function
intersect_sim_obs before calling the criterion functions).
}
