% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_protocol.R
\name{load_protocol_agmip}
\alias{load_protocol_agmip}
\title{Load the protocol description as given in xls file \code{protocol_file_path}}
\usage{
load_protocol_agmip(protocol_file_path, transform_outputs = NULL)
}
\value{
A list including:
}
\description{
Load the protocol description as given in xls file \code{protocol_file_path}
}
\details{
The AgMIP PhaseIV protocol is thoroughly detailed in Wallach et al., 2024
and Wallach et al. 2025.
}
