% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_criteria.R
\name{ls_criteria}
\alias{ls_criteria}
\alias{crit_ols}
\alias{crit_log_cwss}
\alias{crit_log_cwss_corr}
\title{Criteria to optimize}
\usage{
crit_ols(sim_list, obs_list)

crit_log_cwss(sim_list, obs_list)

crit_log_cwss_corr(sim_list, obs_list)
}
\arguments{
\item{sim_list}{List of simulated variables}

\item{obs_list}{List of observed variables}
}
\value{
The value of the criterion given the observed and simulated values of
the variables.
}
\description{
Provide several least squares criteria to estimate parameters by minimizing
the difference between observed and simulated values of model output
variables.
}
\details{
The following criteria are proposed (
\href{https://sticsrpacks.github.io/CroptimizR/reference/ls_criteria.html}{see html version}
for a better rendering of equations):
\itemize{
\item \code{crit_ols}: ordinary least squares \cr
The sum of squared residues for each variable: \cr
\deqn{ \sum_{i,j,k} [Y_{ijk}-f_{jk}(X_i;\theta)]^2 }
where \eqn{ Y_{ijk} } is the observed value for the \eqn{k^{th}} time point of the \eqn{j^{th}} variable in the \eqn{i^{th}}
situation,
\eqn{ f_{jk}(X_i;\theta) } the corresponding model prediction, and \eqn{n_j} the number of measurements of variable \eqn{j}. \cr
In this criterion, one assume that all errors (model and observations errors for all variables, dates and situations) are independent, and that the error variance is constant over time and equal for the different variables \eqn{j}.
\item \code{crit_log_cwss}: log transformation of concentrated version of weighted sum of squares \cr
The concentrated version of weighted sum of squares is: \cr
\deqn{ \prod_{j} {(\frac{1}{n_j} \sum_{i,k} [Y_{ijk}-f_{jk}(X_i;\theta)]^2 )} ^{n_j/2} }
where \eqn{ Y_{ijk} } is the observed value for the \eqn{k^{th}} time point of the \eqn{j^{th}} variable in the \eqn{i^{th}}
situation,
\eqn{ f_{jk}(X_i;\theta) } the corresponding model prediction, and \eqn{n_j} the number of measurements of variable \eqn{j}. \cr
\code{crit_log_cwss} computes the log of this equation. \cr
In this criterion, one assume that all errors (model and observations errors for all variables, dates and situations) are independent, and that the error variance is constant over time but may be different between variables \eqn{j}.
These error variances are automatically estimated.
More details about this criterion are given in Wallach et al. (2011), equation 5.
\item \code{crit_log_cwss_corr}: log transformation of concentrated version of weighted sum of squares with hypothesis of high correlation between errors for different measurements over time \cr
The original criterion is: \cr
\deqn{ \prod_{j} {(\frac{1}{N_j} \sum_{i} [ \frac{1}{n_{ij}} \sum_{k} (Y_{ijk}-f_{jk}(X_i;\theta))^2 ] )} ^{N_j/2} }
where \eqn{ Y_{ijk} } is the observed value for the \eqn{k^{th}} time point of the \eqn{j^{th}} variable in the \eqn{i^{th}}
situation,
\eqn{ f_{jk}(X_i;\theta) } the corresponding model prediction, \eqn{N_j} the number of situations including at least one observation of variable \eqn{j}, and \eqn{n_{ij}} the number of observation of variable \eqn{j} on situation \eqn{i}. . \cr
\code{crit_log_cwss_corr} computes the log of this equation. \cr
In this criterion, one still assume that errors in different
situations or for different variables in the same situation are
independent.
However, errors for different observations over time of the same
variable in the same situation are assumed to be highly correlated.
In this way, each situation contributes a single term to the
overall sum of squared errors regardless of the number of
observations which may be usefull in case one have situations with
very heterogenous number of dates of observations.
More details about this criterion are given in Wallach et al.
(2011), equation 8.
}
\code{sim_list} and \code{obs_list} must have the same structure (i.e. same number of
variables, dates, situations, ... use internal function
intersect_sim_obs before calling the criterion functions).
}
