% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls_criterion.R
\name{ls_criterion}
\alias{ls_criterion}
\alias{crit_cwss}
\alias{crit_log_cwss}
\alias{crit_cwss_corr}
\title{Criterion to optimize}
\usage{
crit_cwss(sim_list, obs_list)

crit_log_cwss(sim_list, obs_list)

crit_cwss_corr(sim_list, obs_list)
}
\arguments{
\item{sim_list}{List of simulated variables}

\item{obs_list}{List of observed variables}
}
\value{
The value of the criterion given the observed and simulated values of the variables.
}
\description{
Provide several least squares criterion to estimate parameters by minimizing the
difference between observed and simulated values of model output variables.
}
\details{
The following criterion are proposed (\href{https://sticsrpacks.github.io/CroptimizR/reference/ls_criterion.html}{see html version} for a better rendering of equations):
\itemize{
\item \code{crit_cwss}: concentrated version of weighted sum of squares
\deqn{ \prod_{j} {(\frac{1}{n_j} \sum_{i,k} [Y_{ijk}-f_{jk}(X_i;\theta)]^2 )} ^{n_j/2} }
where \eqn{ Y_{ijk} } is the observed value for the \eqn{k^{th}} time point of the \eqn{j^{th}} variable in the \eqn{i^{th}}
situation,
\eqn{ f_{jk}(X_i;\theta) } the corresponding model prediction, and \eqn{n_j} the number of measurements of variable \eqn{j}. \cr
More details about this criterion are given in Wallach et al. (2011), equation 5.
\item \code{crit_log_cwss}: log transformation of concentrated version of weighted sum of squares \cr
This criterion can be useful in place of crit_cwss if the sum of residues are null for a given situation
(this may happen for example when one optimize on integers such as phenological stages days ...)
}
\code{sim_list} and \code{obs_list} must have the same structure (i.e. same number of variables, dates, situations, ... use internal function
intersect_sim_obs before calling the criterion functions).
}
