% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_crit.R
\name{main_crit}
\alias{main_crit}
\title{main function for criterion to optimize}
\usage{
main_crit(param_values, crit_options)
}
\arguments{
\item{param_values}{Value(s) of the parameters}

\item{crit_options}{A list containing the following elements:
\code{param_names} Name(s) of parameters
\code{obs_list} List of observed values
\code{crit_function} Function implementing the criterion to optimize
\code{model_function} Function implementing the criterion to optimize
\code{model_options} List of arguments to pass to model function
\code{situation_names} Name(s) of the situations to simulate
\code{prior_information} Prior information on the parameters to estimate.
For the moment only uniform distribution are allowed.
Either a list containing a vector of upper and lower
bounds (\code{ub} and \code{lb}), or a list of list containing for each
parameter and group of situation the names of the situations (\code{sit_names})
and upper and lower bounds (\code{ub} and \code{lb})}
}
\value{
The value of the criterion
}
\description{
main function for criterion to optimize
}
\keyword{internal}
