% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_obsSim_consistent.R
\name{make_obsSim_consistent}
\alias{make_obsSim_consistent}
\title{Check the consistency of observation and simulation lists and make them consistent if possible}
\usage{
make_obsSim_consistent(sim_list, obs_list)
}
\arguments{
\item{sim_list}{List of simulated values}

\item{obs_list}{List of observed values}
}
\value{
A list containing consistent obs and sim lists
}
\description{
Check the consistency of observation and simulation lists and make them consistent if possible
}
\details{
This function checks:
(i) the homogeneity of types of obs and sim Date columns
(ii) that variables in obs and sim list have same types and are numeric
}
\keyword{internal}
