% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_params.R
\name{normalize_params}
\alias{normalize_params}
\title{Normalize parameters}
\usage{
normalize_params(param_info)
}
\arguments{
\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item (named) vectors of upper and lower bounds (\code{ub} and \code{lb}) (-Inf and Inf can be used),
\item \code{init_values}, A data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided), the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or
a named list containing for each parameter the list of situations per group
(\code{sit_list}), the vector of upper and lower bounds (one value per group)
(\code{ub} and \code{lb}) and the list of initial values per group
\code{init_values} (data.frame, one column per group, optional).
(see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example)}
}
\value{
Returns normalized values for param_info
}
\description{
Normalize parameters
}
\details{
first version just for a test ... should include different ways of handling
normalization (from bounds or from given values) + handle the different shapes of param_info ...
}
\examples{


}
\keyword{internal}
