% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_switch.R
\name{optim_switch}
\alias{optim_switch}
\title{Call the required parameter estimation method}
\usage{
optim_switch(
  param_names,
  optim_method,
  optim_options,
  param_info,
  crit_options
)
}
\arguments{
\item{param_names}{Name(s) of parameters to estimate (a parameter name must
be replicated if several groups of situations for this parameter)}

\item{optim_method}{Name of the parameter estimation method to use (optional,
see default value in the function signature). For the moment, can be "simplex"
or "dreamzs". See \href{https://sticsrpacks.github.io/CroptimizR/articles/Available_parameter_estimation_algorithms.html}{here}
for a brief description and references on the available methods.}

\item{optim_options}{List of options of the parameter estimation method, containing:
\itemize{
\item \code{path_results} The path where to store the results (optional, default=getwd())
\item specific options depending on the method used. Click on the links to see examples with the \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_simple_case.html}{simplex}
and \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_DREAM.html}{DreamZS} methods.
}}

\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item \code{ub} and \code{lb}, named vectors of upper and lower bounds (-Inf and Inf can be used if init_values is provided),
\item \code{init_values}, a data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided, the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or a named list containing for each parameter:
\itemize{
\item \code{sit_list}, the list of situations per group,
\item \code{ub} and \code{lb}, vectors of upper and lower bounds (one value per group),
\item \code{init_values}, the list of initial values per group  (data.frame, one column per group, optional).
(see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example)
}}

\item{crit_options}{List containing several arguments given to \code{estim_param} function:
\code{param_names}, \code{obs_list}, \code{crit_function}, \code{model_function}, \code{model_options},
\code{param_info}, \code{transform_obs}, \code{transform_sim}
that must be passed to main_crit function by the methods wrappers.}
}
\value{
prints, graphs and a list containing the results of the parameter estimation,
which content depends on the method used, all that saved in the defined in
\code{optim_options.path_results}
}
\description{
Call the required parameter estimation method
}
\keyword{internal}
