% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bayesian.R
\name{plot_bayesian}
\alias{plot_bayesian}
\title{Generate plots for bayesian methods}
\usage{
plot_bayesian(optim_options, param_info, optim_results)
}
\arguments{
\item{optim_options}{List of options of the parameter estimation method, containing:
\itemize{
\item \code{path_results} The path where to store the results (optional, default=getwd())
\item specific options depending on the method used. Click on the links to see examples with the \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_simple_case.html}{simplex}
and \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_DREAM.html}{DreamZS} methods.
}}

\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item (named) vectors of upper and lower bounds (\code{ub} and \code{lb}) (-Inf and Inf can be used),
\item \code{init_values}, A data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided, the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or
a named list containing for each parameter the list of situations per group
(\code{sit_list}), the vector of upper and lower bounds (one value per group)
(\code{ub} and \code{lb}) and the list of initial values per group
\code{init_values} (data.frame, one column per group, optional).
(see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example)}

\item{optim_results}{Results list returned by bayesian method wrappers}
}
\value{
Save plots in optim_options$path_results.
}
\description{
Generate plots for bayesian methods
}
\keyword{internal}
