% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_frequentist.R
\name{plot_valuesVSit}
\alias{plot_valuesVSit}
\title{Create plots of parameters and criterion values per iteration or evaluation number}
\usage{
plot_valuesVSit(
  df,
  param_info,
  iter_or_eval = c("iter", "eval"),
  crit_log = FALSE,
  rep_label = c("begin_end", "begin", "end")
)
}
\arguments{
\item{df}{Data.frame containing values of parameters (one column per estimated parameter),
criterion (crit column), repetition number (rep), iteration number (iter) and evaluation number (eval)
(similar to params_and_crit). See Details section for comments about the difference between evaluations and iterations.}

\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item \code{ub} and \code{lb}, named vectors of upper and lower bounds (-Inf and Inf can be used if init_values is provided),
\item \code{init_values}, a data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided, the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or a named list containing for each parameter:
\itemize{
\item \code{sit_list}, the list of situations per group,
\item \code{ub} and \code{lb}, vectors of upper and lower bounds (one value per group),
\item \code{init_values}, the list of initial values per group  (data.frame, one column per group, optional).
(see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example)
}}

\item{iter_or_eval}{Values of the x axis: "iter" for iteration number, "eval" for evaluation number}

\item{crit_log}{If TRUE, consider criterion values in log scale}

\item{rep_label}{Indicate if labels for the repetition number must be plotted at both beginning and end of lines ("begin_end"),
only at the beginning ("begin") or only at the end ("end")}
}
\value{
A named list containing one plot per parameter and a plot for the criterion.
}
\description{
Create plots of parameters and criterion values per iteration or evaluation number
}
\details{
Evaluation means evaluation of the criterion from proposed values of the parameters
by the parameter estimation algorithm.
An iteration is reached when an evaluation lead to a better value of the criterion than the previously obtained values.
There are thus more evaluations than iterations. The criterion decreases when
iteration number increases while it is not the case when evaluation number increases.
}
