% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_functions.R
\name{plot_valuesVSit_2D}
\alias{plot_valuesVSit_2D}
\title{Create 2D plots of parameters values evolution per iteration or
evaluation number}
\usage{
plot_valuesVSit_2D(
  df,
  param_info,
  iter_or_eval = c("eval", "iter"),
  fill = c("crit", "rep"),
  crit_log = TRUE,
  lines = FALSE,
  rep_label = c("begin_end", "begin", "end")
)
}
\arguments{
\item{df}{Data.frame containing values of parameters (one column per
estimated parameter), criterion (crit column), repetition number (rep),
iteration number (iter) and evaluation number (eval)
(similar to params_and_crit).
See Details section for comments about the difference between evaluations
and iterations.}

\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item \code{ub} and \code{lb}, named vectors of upper and lower bounds (-Inf and Inf can be used if init_values is provided),
\item \code{default}, named vectors of default values (optional, corresponding parameters are set to these values when the parameter is part of the \code{candidate_param} list and when it is not estimated ; these values are also used as first initial values when the parameters are estimated)
\item \code{init_values}, a data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided, the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or a named list containing for each parameter:
\itemize{
\item \code{sit_list}, list the groups of situations for which the current estimated
parameter must take different values (see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example),
\item \code{ub} and \code{lb}, vectors of upper and lower bounds (one value per group),
\item \code{init_values}, the list of initial values per group  (data.frame, one column per group, optional).
\item \code{default}, vector of default values per group (optional, the parameter is set to its default value when it is part of the \code{candidate_param} list and when it is not estimated ; the default value is also used as first initial value when the parameter is estimated)
}}

\item{iter_or_eval}{"iter" for plotting the values for each iteration,
"eval" for plotting the values for each evaluation}

\item{fill}{If "crit", colours the points and lines in function of the
minimized criterion value, if "rep" colours in function of the
repetition number.}

\item{crit_log}{If TRUE, consider criterion values in log scale}

\item{lines}{If TRUE add lines between points of a same repetition}

\item{rep_label}{Indicate if labels for the repetition number must be plotted
at both beginning and end of lines ("begin_end"), only at the beginning
("begin") or only at the end ("end")}
}
\value{
A list containing one plot per parameter pair.
}
\description{
Create 2D plots of parameters values evolution per iteration or
evaluation number
}
\details{
Evaluation means evaluation of the criterion from proposed values of
the parameters by the parameter estimation algorithm.
An iteration is reached when an evaluation lead to a better value of the
criterion than the previously obtained values.
There are thus more evaluations than iterations. The criterion decreases when
iteration number increases while it is not the case when evaluation number
increases.
}
