% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global_optim_function.R
\name{plot_valuesVSit_go}
\alias{plot_valuesVSit_go}
\title{Create plots of parameters and criterion values per iteration or
evaluation number}
\usage{
plot_valuesVSit_go(
  df,
  param_info,
  iter_or_eval = c("iter", "eval"),
  crit_log = TRUE,
  ind_label = c("begin_end", "begin", "end")
)
}
\arguments{
\item{df}{Data.frame containing values of each individual and iteration parameters values (one column per
estimated parameter), criterion (crit column), individual index (ind),
iteration number (iter) and evaluation number (eval)
(similar to params_and_crit).
See Details section for comments about the difference between evaluations
and iterations.}

\item{param_info}{List describing the parameters to estimate}

\item{iter_or_eval}{Values of the x axis: "iter" for iteration number,
"eval" for evaluation number}

\item{crit_log}{If TRUE, consider criterion values in log scale}

\item{ind_label}{Indicate if labels for the individual number must be
plotted at both beginning and end of lines ("begin_end"),
only at the beginning ("begin") or only at the end ("end")}
}
\value{
A named list containing one plot per parameter and a plot for the
criterion.
}
\description{
Create plots of parameters and criterion values per iteration or
evaluation number
}
\details{
Evaluation means evaluation of the criterion from proposed values of
the parameters by the parameter estimation algorithm.
An iteration is reached when an evaluation lead to a better value of the
criterion than the previously obtained values.
There are thus more evaluations than iterations. The criterion decreases when
iteration number increases while it is not the case when evaluation number
increases.
}
