% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_params.R
\name{sample_params}
\alias{sample_params}
\title{sample values in given distribution}
\usage{
sample_params(bounds_list, n, seed = NULL)
}
\arguments{
\item{bounds_list}{Either a (optionnally named) list containing a vector of
upper and lower bounds (\code{ub} and \code{lb}), or a named list containing
for each parameter the list of situations per group (\code{sit_list})
and the vector of upper and lower bounds (one value per group) (\code{ub} and
\code{lb})}

\item{n}{The number of values to sample for each parameter}

\item{seed}{The seed of the random generator (optional, default value=NULL)}
}
\value{
A vector or data.frame containing the sampled values (nrow=n)
}
\description{
sample values in given distribution
}
\details{
For the moment only works with uniform distributions but will be
hopefully soon extended to any distribution. Use randomized LHS (from
DiceDesign package)
}
\examples{
bounds_list <- list(lb = c(0, 1, 2), ub = c(1, 2, 3))
CroptimizR:::sample_params(bounds_list, 5)
}
\keyword{internal}
