% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FwdRegAgMIP.R
\name{save_results_FwdRegAgMIP}
\alias{save_results_FwdRegAgMIP}
\title{Save the results of the Forward Selection algorithm proposed in
AgMIP calibration phaseIII protocol}
\usage{
save_results_FwdRegAgMIP(param_selection_steps, path_results)
}
\arguments{
\item{param_selection_steps}{A tibble summarizing the results of the previous
parameter estimation steps as returned by the previous call to this function,
NULL if it is the first step.}

\item{path_results}{Path of the folder where to store the results}
}
\value{
Save param_selection_steps in a csv file in folder path_results
}
\description{
Save the results of the Forward Selection algorithm proposed in
AgMIP calibration phaseIII protocol
}
\keyword{internal}
