% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FwdRegAgMIP.R
\name{summary_FwdRegAgMIP}
\alias{summary_FwdRegAgMIP}
\title{Summarize the results of the Forward Selection algorithm proposed in
AgMIP calibration phaseIII protocol}
\usage{
summary_FwdRegAgMIP(param_selection_steps, info_crit_list, path_results)
}
\arguments{
\item{param_selection_steps}{A tibble summarizing the results of the previous
parameter estimation steps as returned by the previous call to this function,
NULL if it is the first step.}
}
\value{
Print the results in standard output.
}
\description{
Summarize the results of the Forward Selection algorithm proposed in
AgMIP calibration phaseIII protocol
}
\keyword{internal}
