% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_steps_functions.R
\name{summary_multi_step}
\alias{summary_multi_step}
\title{Summarizes results of multi-step procedure}
\usage{
summary_multi_step(results_multi_step, path_results)
}
\arguments{
\item{results_multi_step}{Results of the multi_step procedure as returned by post_treat_multi_step}

\item{path_results}{Folder path where results of the multi-step optimization process can be found}
}
\value{
Prints results of the multi-step procedure
}
\description{
Summarizes results of multi-step procedure
}
