% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_wrapper.R
\name{test_wrapper}
\alias{test_wrapper}
\title{Test model wrappers}
\usage{
test_wrapper(
  model_function,
  model_options,
  param_values,
  situation,
  var = NULL,
  sit_names = lifecycle::deprecated(),
  var_names = lifecycle::deprecated()
)
}
\arguments{
\item{model_function}{Crop Model wrapper function to use.}

\item{model_options}{List of options for the Crop Model wrapper (see help of
the Crop Model wrapper function used).}

\item{param_values}{a named vector that contains values and names for AT LEAST TWO model parameters THAT ARE EXPECTED TO PLAY ON ITS RESULTS.}

\item{situation}{Vector of situations names for which results must be tested.}

\item{var}{(optional) Vector of variables names for which results must be tested.}

\item{sit_names}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{sit_names} is no
longer supported, use \code{situation} instead.}

\item{var_names}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{var_names} is no
longer supported, use \code{var} instead.}
}
\value{
A list containing:
\itemize{
\item test_results: a vector of boolean indicating which test succeeded (TRUE) or failed (FALSE)
\item param_values_1: first subset of param_values
\item param_values_2: second subset of param_values
\item sim_1: results obtained with param_values_1
\item sim_2: results obtained with param_values_2
\item sim_3: results obtained for second run with param_values_1
}
}
\description{
This function perform some tests of CroptimizR model wrappers.
See @details for more information.
}
\details{
This function runs the wrapper consecutively with different subsets of param_values.
It then checks:
\itemize{
\item the format of the returned results
\item the results are different when different subsets of param_values are used,
\item the results are identical when same subsets of param_values are used.
}
}
