% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_param.R
\name{validate_step}
\alias{validate_step}
\title{Validate the information for a given step}
\usage{
validate_step(step, istep)
}
\arguments{
\item{step}{A list containing the characteristics of a given step.}

\item{istep}{The index of the step in the list of steps.}
}
\value{
The validated step with possibly updated information.
}
\description{
Validate the information for a given step
}
\keyword{internal}
