% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_param.R
\name{validate_steps}
\alias{validate_steps}
\title{Validate the information for all steps}
\usage{
validate_steps(step_list)
}
\arguments{
\item{step_list}{A list of steps, each containing the characteristics of a given step.}
}
\value{
The validated list of steps with possibly updated information.
}
\description{
Validate the information for all steps
}
\keyword{internal}
