% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_BayesianTools.R
\name{wrap_BayesianTools}
\alias{wrap_BayesianTools}
\title{A wrapper for BayesianTools package}
\usage{
wrap_BayesianTools(optim_options, param_info, crit_options)
}
\value{
prints, graphs and a list containing:
}
\description{
A wrapper for BayesianTools package
}
\keyword{internal}
