% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_parma.R
\name{wrap_cmaes}
\alias{wrap_cmaes}
\title{A wrapper for parma::cmaes package (CMA-ES)}
\usage{
wrap_cmaes(optim_options, param_info, crit_options)
}
\arguments{
\item{optim_options}{see description in estim_param}

\item{param_info}{see description in estim_param}

\item{crit_options}{List containing several arguments given to \code{estim_param}
function: \code{param_names}, \code{obs_list}, \code{crit_function}, \code{model_function},
\code{model_options}, \code{param_info}, \code{transform_obs}, \code{transform_sim}, \code{out_dir}
that must be passed to main_crit function by the methods wrappers.}
}
\value{
list with final_values, init_values, est_values, GE, min_crit_value
}
\description{
A wrapper for parma::cmaes package (CMA-ES)
}
\keyword{internal}
