% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_DE.R
\name{wrap_DEoptim}
\alias{wrap_DEoptim}
\title{A wrapper for DEoptim package}
\usage{
wrap_DEoptim(optim_options, param_info, crit_options)
}
\arguments{
\item{optim_options}{see description in estim_param}

\item{param_info}{see description in estim_param}

\item{crit_options}{List containing several arguments given to \code{estim_param}
function: \code{param_names}, \code{obs_list}, \code{crit_function}, \code{model_function},
\code{model_options}, \code{param_info}, \code{transform_obs}, \code{transform_sim}, \code{out_dir}
that must be passed to main_crit function by the methods wrappers.}
}
\value{
prints, graphs and a list containing:
\code{final_values}, the vector of estimated values for optimized parameters
for the repetition that lead to the lowest value of the criterion
\code{init_values}, the vector of initial values for optimized parameters
\code{est_values}, data frame of values of all population
\code{ind_min_crit}, the index of the repetition that lead to the lowest value
of the criterion
\code{DE}, the data structure returned by DE
\code{crit_values} the vector of criterion values for the all population
}
\description{
A wrapper for DEoptim package
}
\keyword{internal}
