% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_rgenoud.R
\name{wrap_rgenoud}
\alias{wrap_rgenoud}
\title{A wrapper for rgenoud package (single criterion)}
\usage{
wrap_rgenoud(optim_options, param_info, crit_options)
}
\arguments{
\item{optim_options}{see description in estim_param}

\item{param_info}{see description in estim_param}

\item{crit_options}{List containing several arguments given to \code{estim_param}
function: \code{param_names}, \code{obs_list}, \code{crit_function}, \code{model_function},
\code{model_options}, \code{param_info}, \code{transform_obs}, \code{transform_sim}, \code{out_dir}
that must be passed to main_crit function by the methods wrappers.}
}
\value{
A list containing:
\code{final_values}: vector of estimated parameter values
corresponding to the best solution.
\code{init_values}: vector or matrix of initial parameter values.
\code{est_values}: matrix of estimated values (final solution).
\code{min_crit_value}: minimum value of the criterion.
\code{trace_df}: data.frame containing parameters, criterion,
iteration, evaluation and individual indices .
\code{obs_var_list}: list of observed variables used by the criterion.
\code{elapsed}: computation time.
\code{genoud.pro}: an output file called located in the temporary
directory provided by tempdir, containing solutions in each
generation (criterion and parameter values )
}
\description{
A wrapper for rgenoud package (single criterion)
}
\keyword{internal}
