% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_protocol.R
\name{check_protocol_content}
\alias{check_protocol_content}
\title{Check the content of a protocol description file}
\usage{
check_protocol_content(
  protocol_file_path,
  variables_df,
  varNames_corresp,
  simVar_units,
  transform_outputs,
  param_group,
  obsVar_group,
  sitNames_corresp
)
}
\arguments{
\item{variables_df}{}

\item{varNames_corresp}{}

\item{simVar_units}{}

\item{transform_outputs}{}

\item{param_group}{}

\item{obsVar_group}{}

\item{sitNames_corresp}{}
}
\value{
Stop if the content of the protocol description file is not as expected
}
\description{
Check the content of a protocol description file
}
\keyword{internal}
