% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_protocol.R
\name{check_protocol_structure}
\alias{check_protocol_structure}
\title{Check the protocol description file structure}
\usage{
check_protocol_structure(protocol_file_path)
}
\value{
Stop if protocol structure is not as expected
}
\description{
Check the protocol description file structure
}
\details{
Check that the protocol description as given in xls file \code{protocol_file_path}
include the required sheets and columns.
}
\keyword{internal}
