% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_init_values.R
\name{complete_init_values}
\alias{complete_init_values}
\title{Complete given initial values by sampling values from bounds taking into account inequality constraints between parameters}
\usage{
complete_init_values(
  init_values,
  nb_values,
  lb,
  ub,
  ranseed = NULL,
  satisfy_par_const = NULL
)
}
\arguments{
\item{init_values}{A data.frame containing some initial
values for the parameters (or NULL)}

\item{nb_values}{Total number of initial values required for the parameters}

\item{lb}{Lower bounds of the parameters}

\item{ub}{Upper bounds of the parameters}

\item{ranseed}{Set random seed so that each execution give the same results.
If you want randomization, set it to NULL}

\item{satisfy_par_const}{User function for including constraints on estimated
parameters (optional), see details section for more information}
}
\value{
A data.frame containing initial values for all parameters and repetitions
of the minimization process that includes the initial values given in \code{init_values}
plus some randomly chosen ones in bounds \code{lb} and \code{ub}
}
\description{
Complete given initial values by sampling values from bounds taking into account inequality constraints between parameters
}
\examples{

# No user initial values provided nor constraints
lb = c(dlaimax = 0.0005, durvieF = 50)
ub = c(dlaimax = 0.0025, durvieF = 400)
CroptimizR:::complete_init_values(init_values=NULL, nb_values=5, lb=lb, ub=ub)

# Some user initial values are provided but no constraints
lb = c(dlaimax = 0.0005, durvieF = 50)
ub = c(dlaimax = 0.0025, durvieF = 400)
init_values = data.frame(dlaimax=c(0.001, NA), durvieF=c(75, 100))
CroptimizR:::complete_init_values(init_values=NULL, nb_values=5, lb=lb, ub=ub)

}
\keyword{internal}
