% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_param.R
\name{estim_param}
\alias{estim_param}
\title{main function for parameter estimation}
\usage{
estim_param(obs_list, crit_function = crit_log_cwss, model_function,
  model_options = NULL, optim_method = "nloptr.simplex",
  optim_options = list(path_results = getwd(), nmax_NA = 10), param_info,
  transform_obs = NULL, transform_sim = NULL)
}
\arguments{
\item{obs_list}{List of observed values to use for parameter estimation
A \code{named list} (names = situations names) of data.frame containing
one column named Date with the dates (Date or POSIXct format) of the different observations
and one column per observed variables with either the measured values or NA, if
the variable is not observed at the given date.}

\item{crit_function}{Function implementing the criterion to optimize
(optional, see default value in the function signature). See
\href{https://sticsrpacks.github.io/CroptimizR/reference/ls_criterion.html}{here}
for more details about the list of proposed criterion.}

\item{model_function}{Crop Model wrapper function to use.}

\item{model_options}{List of options for the Crop Model wrapper (see help of
the Crop Model wrapper function used).}

\item{optim_method}{Name of the parameter estimation method to use (optional,
see default value in the function signature). For the moment, can be "simplex"
or "dreamzs". See \href{https://sticsrpacks.github.io/CroptimizR/articles/Available_parameter_estimation_algorithms.html}{here}
for a brief description and references on the available methods.}

\item{optim_options}{List of options of the parameter estimation method.
\code{path_results} The path where to store the results (optional, default=getwd())
Click on the links to see the specific options for the \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_simple_case.html}{simplex}
and \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_DREAM.html}{DreamZS} methods.
\code{nmax_NA} In case the optimized criterion returns NA, the parameter estimation
process will stop as soon as the criterion takes the NA value \code{nmax_NA} times
in a row (optional, default=10)}

\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item (named) vectors of upper and lower bounds (\code{ub} and \code{lb}) (-Inf and Inf can be used),
\item \code{init_values}, A data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided), the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or
a named list containing for each parameter the list of situations per group
(\code{sit_list}), the vector of upper and lower bounds (one value per group)
(\code{ub} and \code{lb}) and the list of initial values per group
\code{init_values} (data.frame, one column per group, optional).
(see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here} for an example)}

\item{transform_obs}{Function for transforming observations (optional)}

\item{transform_sim}{Function for transforming simulations (optional)}
}
\value{
prints, graphs and a list containing the results of the parameter estimation,
which content depends on the method used, all that saved in the defined in
\code{optim_options.path_results}
}
\description{
main function for parameter estimation
}
\seealso{
For more detail and examples, see the different vignettes in
\href{https://sticsrpacks.github.io/CroptimizR/}{CroptimizR website}
}
