% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_param.R
\name{estim_param}
\alias{estim_param}
\title{main function for parameter estimation}
\usage{
estim_param(
  obs_list,
  crit_function = crit_log_cwss,
  model_function,
  model_options = NULL,
  optim_method = "nloptr.simplex",
  optim_options,
  param_info,
  forced_param_values = NULL,
  transform_obs = NULL,
  transform_sim = NULL,
  satisfy_par_const = NULL,
  var_names = NULL,
  info_level = 1
)
}
\arguments{
\item{obs_list}{List of observed values to use for parameter estimation
A \verb{named list} (names = situations names) of data.frame containing
one column named Date with the dates (Date or POSIXct format) of the different observations
and one column per observed variables with either the measured values or NA, if
the variable is not observed at the given date.}

\item{crit_function}{Function implementing the criterion to optimize
(optional, see default value in the function signature). See
\href{https://sticsrpacks.github.io/CroptimizR/reference/ls_criteria.html}{here}
for more details about the list of proposed criterion.}

\item{model_function}{Crop Model wrapper function to use.}

\item{model_options}{List of options for the Crop Model wrapper (see help of
the Crop Model wrapper function used).}

\item{optim_method}{Name of the parameter estimation method to use (optional,
see default value in the function signature). For the moment, can be "simplex"
or "dreamzs". See \href{https://sticsrpacks.github.io/CroptimizR/articles/Available_parameter_estimation_algorithms.html}{here}
for a brief description and references on the available methods.}

\item{optim_options}{List of options of the parameter estimation method, containing:
\itemize{
\item \code{path_results} The path where to store the results (optional, default=getwd())
\item specific options depending on the method used. Click on the links to see examples with the \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_simple_case.html}{simplex}
and \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_DREAM.html}{DreamZS} methods.
}}

\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item \code{ub} and \code{lb}, named vectors of upper and lower bounds (-Inf and Inf can be used if init_values is provided),
\item \code{init_values}, a data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided, the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or a named list containing for each parameter:
\itemize{
\item \code{sit_list}, the list of situations per group,
\item \code{ub} and \code{lb}, vectors of upper and lower bounds (one value per group),
\item \code{init_values}, the list of initial values per group  (data.frame, one column per group, optional).
(see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example)
}}

\item{forced_param_values}{Named vector, must contain the values for the model parameters
to force (optional, NULL by default). These values will be transferred to the
model wrapper through its param_values argument so that the given parameters
always take the same values for each model simulation. Should not include values
for estimated parameters (i.e. parameters defined in \code{param_info} argument).}

\item{transform_obs}{User function for transforming observations before each criterion
evaluation (optional), see details section for more information}

\item{transform_sim}{User function for transforming simulations before each criterion
evaluation  (optional), see details section for more information}

\item{satisfy_par_const}{User function for including constraints on estimated
parameters (optional), see details section for more information}

\item{var_names}{(optional) List of variables for which the wrapper must return results.
By default the wrapper is asked to simulate only the observed variables. However,
it may be useful to simulate also other variables, typically when transform_sim
and/or transform_obs functions are used. Note however that it is
active only if
the model_function used handles this argument.}

\item{info_level}{(optional) Integer that controls the level of information returned and stored
by estim_param (in addition to the results automatically provided that depends on the method used).
Higher code give more details.
\itemize{
\item \code{0} to add nothing,
\item \code{1} to add criterion and parameters values, and constraint if satisfy_par_const is provided, for each evaluation
(element params_and_crit in the returned list),
\item \code{2} to add model results, after transformation if transform_sim is provided, and after intersection with observations,
i.e. as used to compute the criterion for each evaluation (element sim_intersect in the returned list),
\item \code{3} to add observations, after transformation if transform_obs is provided, and after intersection with simulations,
i.e. as used to compute the criterion for each evaluation (element obs_intersect in the returned list),
\item \code{4} to add all model wrapper results for each evaluation, and all transformations if transform_sim is provided.
(elements sim and sim_transformed in the returned list),
}}
}
\value{
prints, graphs and a list containing the results of the parameter estimation,
which content depends on the method used and on the values of the \code{info_level} argument.
All results are saved in the folder \code{optim_options$path_results}.
See also the functions \code{\link{plot_valuesVSit}} and \code{\link{plot_valuesVSit_2D}}
for performing additional plots on results of frequentist methods.
}
\description{
main function for parameter estimation
}
\details{
The optional argument \code{transform_obs} must be a function with 4 arguments:
\itemize{
\item model_results: the list of simulated results returned by the mode_wrapper used
\item obs_list: the list of observations as given to estim_param function
\item param_values: a named vector containing the current parameters values proposed
by the estimation algorithm
\item model_options: the list of model options as given to estim_param function
It must return a list of observations (same format as \code{obs_list} argument) that
will be used to compute the criterion to optimize.
}

The optional argument \code{transform_sim} must be a function with 4 arguments:
\itemize{
\item model_results: the list of simulated results returned by the mode_wrapper used
\item obs_list: the list of observations as given to estim_param function
\item param_values: a named vector containing the current parameters values proposed
by the estimation algorithm
\item model_options: the list of model options as given to estim_param function
It must return a list of simulated results (same format as this returned by the model wrapper used)
that will be used to compute the criterion to optimize.
}

The optional argument \code{satisfy_par_const} must be a function with 2 arguments:
\itemize{
\item param_values: a named vector containing the current parameters values proposed
by the estimation algorithm
\item model_options: the list of model options as given to estim_param function
It must return a logical indicating if the parameters values satisfies the constraints
(freely defined by the user in the function body) or not.
}
}
\seealso{
For more details and examples, see the different vignettes in
\href{https://sticsrpacks.github.io/CroptimizR/}{CroptimizR website}
}
