% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params_per_sit.R
\name{get_params_per_sit}
\alias{get_params_per_sit}
\title{Return a list of parameters active for a given situation}
\usage{
get_params_per_sit(sit_groups, situation, param_vec)
}
\arguments{
\item{sit_groups}{Defines the groups of situations for each parameter
A named list containing for each parameter the list of situations per group (\code{sit_list})}

\item{situation}{The name of the situation}

\item{param_vec}{The (named) vector of parameters from which to extract values}
}
\value{
A subvector of \code{param_vec} containing only the values that are
active for the given situation
}
\description{
Return a list of parameters active for a given situation
}
\details{
Names of parameters in \code{param_vec} and \code{sit_groups} should
be the same. If \code{param_vec} is not named, the parameters in \code{param_vec}
will be considered in the same order as in \code{sit_group}
}
\examples{
sg=list(p1=list(sit_list=list(c("sit1","sit2","sit3"),c("sit4","sit5","sit6"))),
        p2=list(sit_list=list(c("sit1","sit2","sit3","sit4","sit5","sit6"))))
vec=c(1,2,3)
CroptimizR:::get_params_per_sit(sg,"sit2",vec)   # should give c(1,3)
CroptimizR:::get_params_per_sit(sg,"sit4",vec)   # should give c(2,3)
names(vec)=c("p2","p1","p1")
CroptimizR:::get_params_per_sit(sg,"sit2",vec)   # should give c(1,2)
CroptimizR:::get_params_per_sit(sg,"sit4",vec)   # should give c(1,3)

}
