% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_concentrated_wss.R
\name{log_concentrated_wss}
\alias{log_concentrated_wss}
\title{Compute log of concentrated weighted sum of squares from simulation and observation list}
\usage{
log_concentrated_wss(sim_list, obs_list)
}
\arguments{
\item{sim_list}{List of simulated variables}

\item{obs_list}{List of observed variables}
}
\value{
The value of the log of the concentrated weighted sum of squares (see Wallach et al. 2011, eq. 5)
}
\description{
Compute log of concentrated weighted sum of squares from simulation and observation list
}
\details{
\code{sim_list} and \code{obs_list} must have the same structure (i.e. same number of variables, dates, situations, ... use intersect_sim_obs for that).
}
