% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obsSim_consistency.R
\name{make_obsSim_consistent}
\alias{make_obsSim_consistent}
\title{Try to make observation and simulation lists consistent if possible}
\usage{
make_obsSim_consistent(sim_list, obs_list)
}
\arguments{
\item{sim_list}{List of simulated values}

\item{obs_list}{List of observed values}
}
\value{
A list containing consistent obs and sim lists
}
\description{
Try to make observation and simulation lists consistent if possible
}
\details{
This function checks and fix the homogeneity of types of obs and sim Date columns
}
\keyword{internal}
