% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim_switch.R
\name{optim_switch}
\alias{optim_switch}
\title{Call the required parameter estimation method}
\usage{
optim_switch(
  param_names,
  obs_list,
  crit_function,
  model_function,
  model_options = NULL,
  optim_method = "nloptr.simplex",
  optim_options = NULL,
  prior_information
)
}
\arguments{
\item{param_names}{Name(s) of parameters to estimate (a parameter name must
be replicated if several groups of situations for this parameter)}

\item{obs_list}{List of observed values}

\item{crit_function}{Function implementing the criterion to optimize}

\item{model_function}{Crop Model wrapper function}

\item{model_options}{List of options for the Crop Model wrapper (optional)}

\item{optim_method}{Name of the parameter estimation method (optional,
simplex is the only one interfaced for the moment)}

\item{optim_options}{List of options of the parameter estimation method:
\code{nb_rep}, the number of repetitions (optional, default=1)
\code{xtol_rel}, the tolerance for the stopping criterion on relative
differences on parameters values between 2 iterations (optional, default=1e-5)
\code{maxeval}, the maximum number of criterion evaluation (optional, default=500)
\code{path_results}, the path where to store the results (optional, default=getwd())}

\item{prior_information}{Prior information on the parameters to estimate.
For the moment only uniform distribution are allowed.
Either
a list containing:
\itemize{
\item (named) vectors of upper and lower bounds (\code{ub} and \code{lb}),
\item \code{init_values}, A (column named) vector or data.frame containing initial
values to test for the parameters (optional, if not provided (or if less values
than number of repetitions of the minimization are provided), the (or part
of the) initial values will be randomly generated using LHS sampling within
parameter bounds.
or
a named list containing for each parameter the list of situations per group
(\code{sit_list}), the vector of upper and lower bounds (one value per group)
(\code{ub} and \code{lb}) and the list of initial values per group
\code{init_values} (column named vector or data.frame per group).
}}
}
\value{
prints, graphs and a list containing the results of the parameter estimation.
which content depends on the method used.
e.g. for Nelder meade simplex in nloptr, this list contains
\code{final_values}, the vector of estimated values for optimized parameters
for the repetition that lead to the lowest value of the criterion
\code{init_values}, the vector of initial values for optimized parameters
\code{est_values}, the vector of estimated values for optimized parameters
for all repetitions
\code{ind_min_crit}, the index of the repetition that lead to the lowest value
of the criterion
\code{nlo}, the data structure returned by nloptr
}
\description{
Call the required parameter estimation method
}
