% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesian_functions.R
\name{plot_bayesian}
\alias{plot_bayesian}
\title{Generate plots for bayesian methods}
\usage{
plot_bayesian(optim_options, param_info, optim_results)
}
\arguments{
\item{optim_options}{List of options of the parameter estimation method, containing:
\itemize{
\item \code{out_dir} Directory path where to write the optimization results (optional, default to \code{getwd()})
\item \code{ranseed} Set random seed so that each execution of estim_param give the same
results when using the same seed. If you want randomization, set it to NULL,
otherwise set it to a number of your choice (e.g. 1234) (optional, default to \code{getwd()})
\item specific options depending on the method used. Click on the links to see examples with the \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_simple_case.html}{simplex}
and \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_DREAM.html}{DreamZS} methods.
\item \code{path_results} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{path_results} is no longer supported, use \code{out_dir} instead.
}}

\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item \code{ub} and \code{lb}, named vectors of upper and lower bounds (-Inf and Inf can be used if init_values is provided),
\item \code{init_values}, a data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided, the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or a named list containing for each parameter:
\itemize{
\item \code{sit_list}, list the groups of situations for which the current estimated
parameter must take different values (see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example),
\item \code{ub} and \code{lb}, vectors of upper and lower bounds (one value per group),
\item \code{init_values}, the list of initial values per group  (data.frame, one column per group, optional).
}}

\item{optim_results}{Results list returned by bayesian method wrappers}
}
\value{
Save plots in optim_options$path_results.
}
\description{
Generate plots for bayesian methods
}
\keyword{internal}
