% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequentist_functions.R
\name{post_treat_frequentist}
\alias{post_treat_frequentist}
\title{Post-treat results of frequentist methods}
\usage{
post_treat_frequentist(optim_options, param_info, optim_results, crit_options)
}
\arguments{
\item{optim_options}{List of options of the parameter estimation method, containing:
\itemize{
\item \code{out_dir} Directory path where to write the optimization results (optional, default to \code{getwd()})
\item \code{ranseed} Set random seed so that each execution of estim_param give the same
results when using the same seed. If you want randomization, set it to NULL,
otherwise set it to a number of your choice (e.g. 1234) (optional, default to \code{getwd()})
\item specific options depending on the method used. Click on the links to see examples with the \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_simple_case.html}{simplex}
and \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_DREAM.html}{DreamZS} methods.
\item \code{path_results} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{path_results} is no longer supported, use \code{out_dir} instead.
}}

\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item \code{ub} and \code{lb}, named vectors of upper and lower bounds (-Inf and Inf can be used if init_values is provided),
\item \code{init_values}, a data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided, the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or a named list containing for each parameter:
\itemize{
\item \code{sit_list}, the list of situations per group,
\item \code{ub} and \code{lb}, vectors of upper and lower bounds (one value per group),
\item \code{init_values}, the list of initial values per group  (data.frame, one column per group, optional).
(see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example)
}}

\item{optim_results}{Results list returned by frequentist method wrappers}

\item{crit_options}{List containing several arguments given to \code{estim_param} function:
\code{param_names}, \code{obs_list}, \code{crit_function}, \code{model_function}, \code{model_options},
\code{param_info}, \code{transform_obs}, \code{transform_sim}
that must be passed to main_crit function by the methods wrappers.}
}
\value{
Updated results of frequentist method
}
\description{
Post-treat results of frequentist methods
}
