% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_steps_functions.R
\name{post_treat_multi_step}
\alias{post_treat_multi_step}
\title{Post-treat results of multi-step procedure}
\usage{
post_treat_multi_step(optim_results_list)
}
\arguments{
\item{optim_results_list}{List of results returned for each step of the parameter estimation process}
}
\value{
List of estimated and forced parameters values
}
\description{
Post-treat results of multi-step procedure
}
