% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_protocol_agmip.R
\name{run_protocol_agmip}
\alias{run_protocol_agmip}
\title{Run the AgMIP PhaseIV protocol}
\usage{
run_protocol_agmip(
  model_function,
  model_options,
  obs_list,
  optim_options,
  protocol_file_path = NULL,
  out_dir = getwd(),
  var_to_simulate = NULL,
  transform_sim = NULL,
  transform_obs = NULL,
  transform_var = NULL,
  step = NULL,
  param_info = NULL
)
}
\arguments{
\item{model_function}{Crop Model wrapper function to use.}

\item{model_options}{List of options for the Crop Model wrapper (see help of
the Crop Model wrapper function used).}

\item{obs_list}{List of observed values to use for parameter estimation.
A \verb{named list} (names = situations names) of data.frame containing
one column named Date with the dates (Date or POSIXct format) of the different observations
and one column per observed variables with either the measured values or NA, if
the variable is not observed at the given date.
See details section for more information on the list of observations actually
used during the parameter estimation process.}

\item{optim_options}{List of options of the parameter estimation method, containing:
\itemize{
\item \code{ranseed} Set random seed so that each execution of estim_param give the same
results when using the same seed. If you want randomization, set it to NULL,
otherwise set it to a number of your choice (e.g. 1234) (optional, default to NULL, which means random seed)
\item specific options depending on the method used. Click on the links to see examples with the \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_simple_case.html}{simplex}
and \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_DREAM.html}{DreamZS} methods.
\item \code{out_dir} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Definition of \code{out_dir} in \code{optim_options} is no longer supported, use the new argument \code{out_dir} of \code{estim_param} instead.
}}

\item{protocol_file_path}{Path to the Excel file describing the AgMIP PhaseIV protocol.}

\item{out_dir}{Path to the directory where the optimization results will be written. (optional, default to \code{getwd()})}

\item{var_to_simulate}{(optional) List of variables for which the model wrapper must return
results.
By default the wrapper is asked to simulate only the observed variables. However,
it may be useful to simulate also other variables, typically when transform_sim
and/or transform_obs functions are used. Note however that it is
active only if the model_function used handles this argument.
If it is the case, and if the obs_var argument is provided, then the list of observations
used will be restricted to the list of variables given in the obs_var argument,
plus the ones possibly computed by the transform_sim function.}

\item{transform_sim}{User function for transforming simulations before each criterion
evaluation  (optional), see details section for more information.}

\item{transform_obs}{User function for transforming observations before each criterion
evaluation (optional), see details section for more information.}

\item{transform_var}{Named vector of functions to apply both on simulated and
observed variables. \code{transform_var=c(var1=log, var2=sqrt)} will for example
apply log-transformation on simulated and observed values of variable var1,
and square-root transformation on values of variable var2.}

\item{step}{(optional) List that describes the steps of the parameter estimation process (see details section).
If \code{NULL}, a single default step will be created using the \code{estim_param} arguments}

\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item \code{ub} and \code{lb}, named vectors of upper and lower bounds (-Inf and Inf can be used if init_values is provided),
\item \code{default}, named vectors of default values (optional, corresponding parameters are set to these values when the parameter is part of the \code{candidate_param} list and when it is not estimated ; these values are also used as first initial values when the parameters are estimated)
\item \code{init_values}, a data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided, the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or a named list containing for each parameter:
\itemize{
\item \code{sit_list}, list the groups of situations for which the current estimated
parameter must take different values (see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example),
\item \code{ub} and \code{lb}, vectors of upper and lower bounds (one value per group),
\item \code{init_values}, the list of initial values per group  (data.frame, one column per group, optional).
\item \code{default}, vector of default values per group (optional, the parameter is set to its default value when it is part of the \code{candidate_param} list and when it is not estimated ; the default value is also used as first initial value when the parameter is estimated)
}}
}
\value{
prints, graphs and a list containing the results of the AgMIP PhaseIV protocol.
All results are saved in the folder \code{out_dir}.
}
\description{
Run the AgMIP PhaseIV protocol
}
\details{
The AgMIP PhaseIV protocol is thoroughly detailed in Wallach et al. (2024)
and Wallach et al. (2025).
}
\seealso{
Wallach et al. (2024), Wallach et al. (2025), and the \code{estim_param} function.
}
