% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_info_functions.R
\name{set_init_values}
\alias{set_init_values}
\title{Set initial values in param_info}
\usage{
set_init_values(param_info, init_values)
}
\arguments{
\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item \code{ub} and \code{lb}, named vectors of upper and lower bounds (-Inf and Inf can be used if init_values is provided),
\item \code{init_values}, a data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided, the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or a named list containing for each parameter:
\itemize{
\item \code{sit_list}, list the groups of situations for which the current estimated
parameter must take different values (see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here}
for an example),
\item \code{ub} and \code{lb}, vectors of upper and lower bounds (one value per group),
\item \code{init_values}, the list of initial values per group  (data.frame, one column per group, optional).
}}

\item{init_values}{A data.frame containing the initial values to set in param_info
(one column per parameter)}
}
\value{
A list similar to param_info but with the new initial values
}
\description{
Set initial values in param_info
}
\keyword{internal}
