% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FwdRegAgMIP.R
\name{summary_FwdRegAgMIP}
\alias{summary_FwdRegAgMIP}
\title{Summarize the results of the Forward Selection algorithm proposed in
AgMIP calibration phaseIII protocol}
\usage{
summary_FwdRegAgMIP(
  param_selection_steps,
  info_crit_list,
  path_results,
  optim_results,
  indent = 0
)
}
\arguments{
\item{param_selection_steps}{A tibble summarizing the results of the
parameter estimation steps, as returned by the post_treat_FwdRegAgMIP function.}

\item{optim_results}{Results list returned by the parameter estimation method
wrapper used}

\item{indent}{Integer, level of indent of the printed messages as required by make_display_prefix}
}
\value{
Print the results in standard output.
}
\description{
Summarize the results of the Forward Selection algorithm proposed in
AgMIP calibration phaseIII protocol
}
\keyword{internal}
