% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estim_param.R
\name{validate_step}
\alias{validate_step}
\title{Validate the information for a given step}
\usage{
validate_step(step, step_name)
}
\arguments{
\item{step}{A list containing the characteristics of a given step.}

\item{step_name}{The name of the step.}
}
\value{
The validated step with possibly updated information.
}
\description{
Validate the information for a given step
}
\keyword{internal}
