% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_BayesianTools.R
\name{wrap_BayesianTools}
\alias{wrap_BayesianTools}
\title{A wrapper for BayesianTools package}
\usage{
wrap_BayesianTools(param_names, optim_options, param_info, crit_options)
}
\arguments{
\item{param_names}{Name(s) of parameters to estimate (a parameter name must
be replicated if several groups of situations for this parameter)}

\item{optim_options}{List of options of the parameter estimation method.
\code{path_results} The path where to store the results (optional, default=getwd())
Click on the links to see the specific options for the \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_simple_case.html}{simplex}
and \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_DREAM.html}{DreamZS} methods.
\code{nmax_NA} In case the optimized criterion returns NA, the parameter estimation
process will stop as soon as the criterion takes the NA value \code{nmax_NA} times
in a row (optional, default=10)}

\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item (named) vectors of upper and lower bounds (\code{ub} and \code{lb}) (-Inf and Inf can be used),
\item \code{init_values}, A data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided), the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or
a named list containing for each parameter the list of situations per group
(\code{sit_list}), the vector of upper and lower bounds (one value per group)
(\code{ub} and \code{lb}) and the list of initial values per group
\code{init_values} (data.frame, one column per group, optional).
(see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here} for an example)}

\item{crit_options}{List containing several arguments given to \code{estim_param} function:
\code{param_names}, \code{obs_list}, \code{crit_function}, \code{model_function}, \code{model_options},
\code{param_info}, \code{transform_obs}, \code{transform_sim}
that must be passed to main_crit function by the methods wrappers.}
}
\value{
prints, graphs and a list containing:
}
\description{
A wrapper for BayesianTools package
}
\keyword{internal}
