% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_BayesianTools.R
\name{wrap_BayesianTools}
\alias{wrap_BayesianTools}
\title{A wrapper for BayesianTools package}
\usage{
wrap_BayesianTools(optim_options, param_info, crit_options)
}
\arguments{
\item{optim_options}{see description in estim_param}

\item{param_info}{see description in estim_param}

\item{crit_options}{List containing several arguments given to \code{estim_param}
function: \code{param_names}, \code{obs_list}, \code{crit_function}, \code{model_function},
\code{model_options}, \code{param_info}, \code{transform_obs}, \code{transform_sim}, \code{out_dir}
that must be passed to main_crit function by the methods wrappers.}
}
\value{
prints, graphs and a list containing:
}
\description{
A wrapper for BayesianTools package
}
\keyword{internal}
