% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_nloptr.R
\name{wrap_nloptr}
\alias{wrap_nloptr}
\title{A wrapper for nloptr package}
\usage{
wrap_nloptr(optim_options, param_info, crit_options)
}
\value{
prints, graphs and a list containing:
\code{final_values}, the vector of estimated values for optimized parameters
for the repetition that lead to the lowest value of the criterion
\code{init_values}, the vector of initial values for optimized parameters
\code{est_values}, the vector of estimated values for optimized parameters
for all repetitions
\code{ind_min_crit}, the index of the repetition that lead to the lowest value
of the criterion
\code{nlo}, the data structure returned by nloptr
}
\description{
A wrapper for nloptr package
}
\keyword{internal}
