% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_nloptr.R
\name{wrap_nloptr}
\alias{wrap_nloptr}
\title{A wrapper for nloptr package}
\usage{
wrap_nloptr(param_names, obs_list, crit_function, model_function,
  model_options = NULL, optim_options = NULL, param_info)
}
\arguments{
\item{param_names}{Name(s) of parameters to estimate (a parameter name must
be replicated if several groups of situations for this parameter)}

\item{obs_list}{List of observed values to use for parameter estimation
A \code{named list} (names = situations names) of data.frame containing
one column named Date with the dates (Date or POSIXct format) of the different observations
and one column per observed variables with either the measured values or NA, if
the variable is not observed at the given date.}

\item{crit_function}{Function implementing the criterion to optimize
(optional, see default value in the function signature). See
\href{https://sticsrpacks.github.io/CroptimizR/reference/ls_criterion.html}{here}
for more details about the list of proposed criterion.}

\item{model_function}{Crop Model wrapper function to use.}

\item{model_options}{List of options for the Crop Model wrapper (see help of
the Crop Model wrapper function used).}

\item{optim_options}{List of options of the parameter estimation method.
\code{path_results} The path where to store the results (optional, default=getwd())
Click on the links to see the specific options for the \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_simple_case.html}{simplex}
and \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_DREAM.html}{DreamZS} methods.}

\item{param_info}{Information on the parameters to estimate.
Either
a list containing:
\itemize{
\item (named) vectors of upper and lower bounds (\code{ub} and \code{lb}) (-Inf and Inf can be used),
\item \code{init_values}, A data.frame containing initial
values to test for the parameters (optional, if not provided, or if less values
than number of repetitions of the minimization are provided), the, or part
of the, initial values will be randomly generated using LHS sampling within
parameter bounds).
}

or
a named list containing for each parameter the list of situations per group
(\code{sit_list}), the vector of upper and lower bounds (one value per group)
(\code{ub} and \code{lb}) and the list of initial values per group
\code{init_values} (data.frame, one column per group, optional).
(see \href{https://sticsrpacks.github.io/CroptimizR/articles/Parameter_estimation_Specific_and_Varietal.html}{here} for an example)}
}
\value{
prints, graphs and a list containing:
\code{final_values}, the vector of estimated values for optimized parameters
for the repetition that lead to the lowest value of the criterion
\code{init_values}, the vector of initial values for optimized parameters
\code{est_values}, the vector of estimated values for optimized parameters
for all repetitions
\code{ind_min_crit}, the index of the repetition that lead to the lowest value
of the criterion
\code{nlo}, the data structure returned by nloptr
}
\description{
A wrapper for nloptr package
}
\keyword{internal}
